update on Qa

+++ apollo-store/src/main/kotlin/org/kebs/app/kotlin/apollo/store/model/qa/QaProcessStatusEntity.kt
@@ -24,7 +24,7 @@ class QaProcessStatusEntity : Serializable {
 
     @Column(name = "STATUS")
     @Basic
-    var status: Long? = null
+    var status: Int? = null
 
     @Column(name = "VAR_FIELD_1")
     @Basic
